Split Settings class finctionality
- Register: holds code to tight functionality to Wordpress 
- Screen: : holds the code to display the Settings screen in the Admin area
- Data: interacts the settings Data
